Populate Recipe for Fettucine Alfredo
On the fettucine-alfredo.html page, I added a photo, photo credits,
description, ingredients, and steps for creating fettucine alfredo.
I used an unordered list to create the ingredients since the order
does not matter. I used an ordered list for the steps since the order
matters. The contents for the Fettucine Alfredo page is now complete.