Resolve hyperlink on beef bulgogi page
The hyperlink for the homepage on the beef bulgogi page was missing a
closing tag, causing the entire rest of the text on the page to be a
hyperlink. The page now only has a one-word hyperlink at the top of the
page.  Additionally, I resized the beef bulgogi image to better match
the size of the following text.